Cleaned up package name
