+++ src/main/java/cashpiles/model/Account.java
@@ -1,4 +1,4 @@
-package cashpiles.model2;
+package cashpiles.model;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
