+++ src/main/java/cashpiles/model/AccountsView.java
@@ -1,4 +1,4 @@
-package cashpiles.model2;
+package cashpiles.model;
 
 import cashpiles.currency.Amount;
 import cashpiles.ledger.AccountCommand;
