+++ src/main/java/cashpiles/model/BudgetPeriod.java
@@ -1,4 +1,4 @@
-package cashpiles.model2;
+package cashpiles.model;
 
 import java.util.List;
 import java.util.TreeMap;
