+++ src/main/java/cashpiles/model/IncomeAllocator.java
@@ -1,4 +1,4 @@
-package cashpiles.model2;
+package cashpiles.model;
 
 import cashpiles.currency.Amount;
