+++ src/main/java/cashpiles/model/Ledger.java
@@ -1,4 +1,4 @@
-package cashpiles.model2;
+package cashpiles.model;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
