+++ src/main/java/cashpiles/model/ManualGoalCategory.java
@@ -1,4 +1,4 @@
-package cashpiles.model2;
+package cashpiles.model;
 
 import cashpiles.ledger.ManualGoalBudgetEntry;
